Basic BigInt support
We should not crash when passed a bigint. This patch will make it so that
the formatter will convert the bigint to a number if possible, else it will
emit an overflow string (#####). A setting has been added so that instead
of the overflow string, a string version of the bigint is emitted instead.
This is similar how date overflows are handled.

I have made an attempt at supporting bigint throughout the library with
some success, but I have abandoned that because it was complicating the
code a lot and I had some worries about performance.

I think partial support would be possible with low cost. But I don't
have the resources to spare for that right now.

Closes #61